requirements.txt
google-generativeai>=0.3.0
openai>=1.0.0
python-dotenv>=1.0.0
google-genai
coverage
mutmut
datasets
tqdm
python-dotenv
# unittest